requirements.txt
django-allauth
django-crispy-forms-ng
django-extensions
django-taggit-templatetags
django-taggit-labels
django-taggit-autosuggest
django-grappelli
django
